Split bots code from server
